Address sidebar ExtensionKit review feedback
