+++ Packages/CMUXExtensionClient/README.md
@@ -2,10 +2,12 @@
 
 `CMUXExtensionClient` is the host-side package for loading and driving CMUX sidebar extensions.
 
-This first slice contains:
+This package contains:
 
 - manifest validation through `CmuxExtensionKit`
 - a registry for available sidebar extensions
 - a session actor that fetches snapshots and dispatches sidebar actions
+- third-party sidebar extension discovery for `com.manaflow.cmux.sidebar`
+- a SwiftUI `EXHostViewController` bridge for rendering the extension scene
 
-The package is ready for an ExtensionKit-backed adapter, but this branch keeps the first verified surface focused on the API and host lifecycle. Future work should add the real `EXHostViewController` bridge and third-party extension discovery.
+The exported surface is intentionally sidebar-only: `CmuxExtensionKit` defines the manifest, snapshot, and action contract, while this package owns host lifecycle, discovery, registry lookup, session dispatch, and the ExtensionKit host view.
